Add author info and address

+++ dcsfinal.py
@@ -2,17 +2,19 @@ from Bio import SeqIO
 from Bio.Blast import NCBIWWW
 from Bio.Blast import NCBIXML
 from progress.bar import Bar
+import os
+import re
 
+if not (os.path.exists("results.xml")):
+    sequence_file = input("Enter the name of your sequence file in FASTA format:")
+    sequence_data = open(sequence_file).read()
+    bar = Bar("Processing sequence data... ")
+    result_handle = NCBIWWW.qblast("blastn", "nt", sequence_data)
+    bar.finish()
 
-sequence_file = input("Enter the name of your sequence file in FASTA format: ")
-sequence_data = open(sequence_file).read()
-bar = Bar("Processing sequence data: ")
-result_handle = NCBIWWW.qblast("blastn", "nt", sequence_data)
-bar.finish()
-
-with open("results.xml","w") as save_file:
-    blast_results = result_handle.read()
-    save_file.write(blast_results)
+    with open("results.xml","w") as save_file:
+        blast_results = result_handle.read()
+        save_file.write(blast_results)
 
 results = open("results.xml","r")
 records = list(NCBIXML.parse(results)) #parse results (xml file) into a list
@@ -28,7 +30,9 @@ for i in range(len(one_query.alignments)): #loop through every query to obtain t
 #parsing GenBank data
 from Bio import Entrez
 Entrez.email = input("Enter your email address (so that NCBI can contact you if there's a problem):")
+#bar = Bar("Parsing location for each accession numbers from GenBank... ", max = len(accession_numbers))
 
+location_list = []
 for i in accession_numbers:
     handle = Entrez.efetch(db = 'nucleotide', rettype = 'gb', retmode = 'text', id = i)
     record = handle.read()
@@ -60,4 +64,63 @@ for i in accession_numbers:
         updated_location = location.split("URL")
         location = updated_location[0]
 
-    print(f"The location where the journal was published for accession number:{i} is {location}")
+    # parse author info
+    author_line = [a for a in info if a.startswith('REFERENCE   2') or a.startswith('  TITLE     Direct Submission')]
+    author_line_indices = []
+    for q in author_line:
+        index_number_ = info.index(q)
+        author_line_indices.append(index_number_)
+
+    index_reference = author_line_indices[0]
+    index_author = index_reference + 1
+    index_title = author_line_indices[-1]
+    author_info = info[index_author:index_title]
+
+    try:
+        first_author_line = author_info[0]
+        first_author_split = first_author_line.split("THORS   ")
+        cleaned_first_author_line = first_author_split[1]
+        author = cleaned_first_author_line
+
+        for k in author_info[1:]:
+            cleaned_line = k.lstrip(" ")
+            author += cleaned_line
+
+
+    except IndexError as error:
+        author_line_indices = []
+        author_line = [g for g in info if g.startswith('  AUTHORS') or g.startswith('  TITLE     Direct Submission')]
+        for l in author_line:
+            index_number_ = info.index(l)
+            author_line_indices.append(index_number_)
+
+        index_author = author_line_indices[0]
+        index_title = author_line_indices[-1]
+        author_info = info[index_author:index_title]
+
+        first_author_line = author_info[0]
+        first_author_split = first_author_line.split("THORS   ")
+        cleaned_first_author_line = first_author_split[1]
+        author = cleaned_first_author_line
+
+        for k in author_info[1:]:
+            cleaned_line = k.lstrip(" ")
+            author += cleaned_line
+
+
+
+    print(f"For accession number {i}, the journal was published by {author} from {location}")
+    location_list.append(location)
+
+    #bar.next()
+#bar.finish()
+
+
+address_line = []
+regex_ = "([\sa-zA-Z0-9-]+,[\sa-zA-Z]+[\s0-9a-zA-Z-]+,[a-zA-Z\s]+)$"
+for i in location_list:
+    result = re.findall(regex_, i)
+    final_result = " ".join(result)
+    address_line.append(final_result.strip())
+
+print(address_line)
